Remove unused functions. Also move all utility functions to a single file.
