+++ interactive_templates/templates/v2/analysis/event_counts.py
@@ -4,8 +4,12 @@ from pathlib import Path
 
 import numpy as np
 import pandas as pd
-from report_utils import get_date_input_file, match_input_files, save_to_json
-from utilities import drop_zero_practices
+from report_utils import (
+    drop_zero_practices,
+    get_date_input_file,
+    match_input_files,
+    save_to_json,
+)
 
 
 def round_to_nearest_100(x):
